// lib/bank.d.ts
// Generated by https://quicktype.io

export interface BankResponse {
  id: string;
  name: string;
  license: number;
  url: string;
  modal: boolean;
  price: string;
  price_int: number;
  price_percent: string;
  discount: string;
  discount_value: string;
  discount_price_int: number;
  discount_price: string;
  discount_price_percent: string;
  pro_discount: string;
  pro_discount_value: string;
  pro_price: string;
  pro_price_int: number;
  pro_price_percent: string;
  rating: string;
}